requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wikipedia==1.4.0
